+++ src/main/java/com/nppgks/reports/service/ReportNameService.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface ReportNameService {
 
-    List<ReportName> getReportNameByDateAndReportId(Long reportTypeId, String dtCreationStart);
+    List<ReportName> getReportNameByDateAndReportId(Integer reportTypeId, String dtCreationStart);
 
     List<ReportName> findAll();
 
-    List<ReportName> findByReportTypeId(Long reportTypeId);
+    List<ReportName> findByReportTypeId(Integer reportTypeId);
 }
